Hush, Drupal.CSS.ClassDefinitionNameSpacing
Suppose you have a CSS directive like this:

  .foo .bar:not(.apple,.banana,.cherry,.date) {
    yaddayadda: 1px;
  }

This rule forces you to split it out like:

  .foo .bar:not(.apple,
  .banana,
  .cherry,
  .date) {
    yaddayadda: 1px;
  }

It basically prohibits you from doing any kind of representation of the
parenthetical that is easy to skim.  On skimming, it wrongly looks like
`.banana` is an entirely separate expression.

So let's just disble this rule...